requirements.txt
scikit-image
torchvision
torch
pillow
numpy 
pandas
scipy 
prdc
openai
requests
facenet-pytorch
bert-score
nltk
torchmetrics
torch-fidelity
ultralytics
torchvision
opencv-python
sentence-transformers
transformers
diffusers
open-clip-torch
shortuuid
bs4
selenium
black